python/requirements.txt
Flask==0.12.2
pyserial==3.3
